requirements.txt
Pillow==12.1.1
playwright==1.58.0
PyMuPDF==1.27.2.2
PyYAML==6.0.2
